Correct array shape after collecting coeffs along the first dimension
Permutedims!(...) was supposed to only shift the last index to the first 
place. Previously, it incorrectly reversed the ordering of the indices 
(essentially giving a transposed FFT in the end).
The unit tests missed this since they were also written in a row-major 
order.
The tests now correctly test in a memory efficient way (column-major 
order) and some have been reduced to a single line.